requirements.txt
pyfiglet==0.8.post1
PyYAML==6.0.2
rich==12.6.0